WIP add Runtime calls to book
