+++ subxt/src/book/usage/runtime_apis.rs
@@ -5,7 +5,34 @@
 //! # Runtime API interface
 //!
 //! The Runtime API interface allows Subxt to call runtime APIs exposed by certain pallets in order
-//! to obtain information.
+//! to obtain information. Much like [`super::storage`] and [`super::extrinsics`], Making a runtime
+//! call to a node and getting the response back takes the following steps:
+//!
+//! 1. [Constructing a runtime call](#constructing-a-runtime-call)
+//! 2. [Submitting it to get back the response](#submitting-it)
+//!
+//! ## Constructing a runtime call
+//!
+//! We can use the statically generated interface to build runtime calls:
+//!
+//! ```rust,no_run
+//! use sp_keyring::AccountKeyring;
+//!
+//! #[subxt::subxt(runtime_metadata_path = "../artifacts/polkadot_metadata.scale")]
+//! pub mod polkadot {}
+//!
+//! let runtime_call = polkadot::apis().metadata().metadata_versions();
+//! ```
+//!
+//! Alternately, we can dynamically construct a runtime call:
+//!
+//! ```rust,no_run
+//! use sp_keyring::AccountKeyring;
+//! use subxt::dynamic::Value;
+//!
+//! let account = AccountKeyring::Alice.to_account_id();
+//! let storage_query = subxt::dynamic::runtime_api_call("Metadata_metadata_versions", vec![], None)
+//! ```
 //!
 //! At the moment, this interface is simply a wrapper around the `state_call` RPC method. This means
 //! that you need to know which runtime calls are available and how to encode their parameters (if
